Added color remover
Added color remover and changed some code
