+++ src/colorBar.ts
